Refactored reducer and seperated functions into meaningful reducers for each type of state.
